third_party/WebKit/LayoutTests/external/wpt/web-animations/interfaces/KeyframeEffect/spacing-expected.txt

This is a testharness.js-based test.
FAIL Test throwing TypeError if using empty string assert_throws: function "function () {
    anim.effect.spacing = '';
  }" did not throw
FAIL Test throwing TypeError if not using the correct keyword assert_throws: function "function () {
    anim.effect.spacing = 'dist';
  }" did not throw
FAIL Test falling back to distribute spacing if using a unrecognized property assert_equals: spacing mode expected "distribute" but got "paced(A)"
FAIL Test falling back to distribute spacing if using CSS variables assert_equals: spacing mode expected "distribute" but got "paced(--bg-color)"
FAIL Test falling back to distribute spacing if using a non-animatable property assert_equals: spacing mode expected "distribute" but got "paced(animation-duration)"
PASS Test spacing value if setting distribute 
PASS Test spacing value if setting paced 
Harness: the test ran to completion.
